# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Platform API (Marble / TaskStack suite)
  description: |
    HTTP JSON API under `/api/v1` for the Marble / TaskStack suite platform service (implemented in this repository's `api` Go module).
    Success bodies use `{ "data": ..., "meta": { "request_id": "..." } }`.
    Errors use `{ "error": { "code", "message", "details?" }, "meta": { "request_id" } }`.

    **Authentication**
    - **Bearer**: HS256 platform JWT (`POST /auth/token` client_credentials) or OIDC access token when `OIDC_*` is configured.
    - **Session cookie**: HttpOnly session from `POST /auth/login`. Mutating requests with cookies must send the CSRF header (`X-CSRF-Token` by default) matching the non-HttpOnly CSRF cookie. Bearer-authenticated mutating requests skip CSRF.

    **Human-only operations**: `PUT` and `DELETE` on `/users/{id}` require an authenticated **end-user** subject (`user:<id>` from session or user-scoped Bearer). Subjects `client:*` (machine / client_credentials) receive 403.

...

    **Economy ledger (read-only):** `GET /economy/ledger` requires permission `economy.read` and a human subject. Query filters: `limit` (1–100, default 50), optional `before_id`, `platform_user_id`, `event_type`, `from`, `to` (RFC3339).

...

...
  version: 0.1.0
servers:
  - url: /
    description: API root (same host as backend; admin SPA may proxy `/api`).
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Login (sets session + CSRF cookies)
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeLoginResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 platform access token or OIDC access token (when configured).
    sessionCookie:
      type: apiKey
      in: cookie
      name: gl_session
      description: Session cookie from login (default name `gl_session`; override with env `SESSION_COOKIE_NAME`). Pair with CSRF cookie + header on unsafe methods.
...
  responses:
    ValidationError:
      description: Bad request / validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: VALIDATION_ERROR
              message: invalid request body
              details:
                field: body
            meta:
              request_id: ""
    UnauthorizedError:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RateLimitedError:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    Meta:
      type: object
      required: [request_id]
      properties:
        request_id:
          type: string
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - UNAUTHORIZED
                - FORBIDDEN
                - CONFLICT
                - INTERNAL_ERROR
                - NOT_IMPLEMENTED
                - RATE_LIMITED
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        meta:
          $ref: "#/components/schemas/Meta"
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    EnvelopeLoginResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [user_id, email]
          properties:
            user_id:
              type: integer
            email:
              type: string
        meta:
          $ref: "#/components/schemas/Meta"
...